Rename `SetPreSignatureDecoder` and usage to `GPv2Decoder`
