+++ src/domain/swaps/contracts/__tests__/encoders/gp-v2-settlement-encoder.builder.ts
@@ -2,7 +2,7 @@ import { faker } from '@faker-js/faker';
 import { encodeFunctionData, Hex, keccak256, toBytes } from 'viem';
 import { Builder } from '@/__tests__/builder';
 import { IEncoder } from '@/__tests__/encoder-builder';
-import { abi } from '@/domain/swaps/contracts/decoders/set-pre-signature-decoder.helper';
+import { abi } from '@/domain/swaps/contracts/decoders/gp-v2-decoder.helper';
 
 type SetPreSignatureArgs = {
   orderUid: `0x${string}`;
